introduce a concept of a reactor 'leader'
in multimaster hot-hot environments, all events are streamed to all masters for
a given minion. this introduces the problem of a reactor being executed on each
hot master, regardless of if the reaction is idempotent  or not. This patch
introduces a simple is_leader bool that can be controlled via the reactor
runner (or emitting the equivalent event to toggle) whether a reactor in
hot-standby mode should be reacting to events coming into the bus. The feature
itself is disconnected from multimaster, and is by default True, so it should
be inert in most saltstack installs. Tests forthcoming in following commit