minion: avoid blocking dns resolution on connect
the salt.utils.network.dns_check function used for resolution also
includes a blocking socket.connect to attempt to confirm the service on
the other side. This is a problem in salt.minion, as it's used as part
of eval_master continually to evaluate the master list; when a master
entry isn't alive / reachable, the socket.connect timeout will block the
minion entirely. This patch turns off that socket.connect for all
eval_master dns_check calls.