Add master_return_strategy feature
This option controls the master return strategy. Can be `source` or
`any`. If set to `source`, then the minion will only attempt to return
the job results to the master that sent the job. If set to `any`, then
the minion will first attempt to return the job results to the master
that sent the job but if that fails, then the minion will attempt to
return the job results to the other configured masters until successful
or the master list has been exhausted. `master_alive_interval` must
also be set. This is used to keep track of which masters are currently
connected.