WIP: recursive ACL at lazyloader level
This is an unfinished proof of concept of how to approach enforcing
authorization not just at the entrypoints of salt (api, master), but
within these calls as well. Since all state calls and native loader
modules often times invoke many other functions, enforcing acl at the
loader level is a more comprehensive attempt at closing that hole and
regulating what a user is and is not able to do. This of course isn't
perfect, but is a step in the direction of fully enforcing an acl.

salt.utils.ctx.RequestContext is used as the mechanism of sharing what
'auth_list' profile is currenly active for a given execution, and is
passed down from master to minion for minion side authorization
enforcement as well.

An uncomprehensive list of things that still need to be done:
 - a concept of master/minion side differentation for allowing code to
   be run on a minion, but NOT on a master (think cmd.run, allowing on
   master gives full root)
 - implementation of other clients (wheel, local, ssh if upstream).
   Currenlty only runner is implemented but others should be
   straightforward to add)
 - Automated tests, but assertions in general that we can rely on
   ContextRequest how we are using it.
 - the loader_acl config opt needs to be added properly to config.py,
   and possibly given a better more obvious name
 - probably other things im forgetting