first cut: bloomberg rhst environments ext_pillar
this is a preliminary cut of rhst tags being used to drive salt
environments. I've fleshed out how to add a SOR based fallback vs
bbcpu.lst which is what hostinfo relies on; but I don't think that is
actually necessary - any provisioning time salt users are global
(lifecycle), so we don't need to care what tags the machine that is not
yet in bbcpu.lst has.
@nsmith269 - we'll need to update salt-repo-sync to map to this slightly
changed schema.