address race condition in LocalClient
The current flow through localclient to Publisher requires waiting for
a response (which contains the jid) before subscribing to the job rets.
There is a race condition in which the publish and response could be
sent before return & subscribe, and there is no reason we can't mint the
jid _before_ publishing and circumvent the issue entirely, so we do
that. The existing master-side code path will still exist for back
compat if a jid is not pre-minted.