improve long polling performance / subscription tracking
when dealing with long running jobs via cmd_iter, the heartbeat find_job
calls would continually add new event.subscribe() tag match
subscriptions for the jobs as they come in. This works out to a slow
leak over time when dealing with a job that runs over 12 hours+, and
adds delays when the real job ret finally does come in. This improves
the subscription handling to unsub the previous saltutil.find_job calls
events if set before broadcasting the next.