salt.auth.runas: new auth backend
This auth backend is a special case that is only meant to be used with
the 'eauth_acl_module' module setting. If a given user is authorized in
the runas eauth configuration to be allowed to run as someone else, the
acl for a given specified user in eauth_opts is returned instead of what
is configured for the given user.