fix runas eauth/eauth_opts in all localclient/runnerclient handling
in order to  correctly runas in publish and runner scnearios some work
needed to be done. Runners now always go through the master inferface to
trigger jobs, never forking and running locally. This has broken one
thing I've noticed, while not strictly broken per-se, jobs like
state.run that rely on emitting to stdout/stderr during execution are
not returned until the job is _complete_, so they don't work as
expected. I'll have to revisit this to see if I can think of
a compromise.