new feature: external_auth default config
when specifying a set of matchers/authorizations in the 'default'
category in external_auth, this will be applied to any authenticated
user. This can be used for common sane things that any user would need
to interact with salt in a normal capacity (ie config.* and others are
implicitly required in any low exec).