teach cache.pgjsonb about its materialized view
there is an implicit assumption that enabling the pgjsonb cache means
that we are using the pgjsonbckminions subclass for salt.util.minions
minion operations with minion data cache. This means that we have to
perform a periodic refresh of the matieralized view that offers fast
access to the ipv4 ip addresses of every minion in our cache. This adds
a simple generalized maintenance interface to account for that, which
will be a simple noop on cache backends that do not support it.