pillar: implement acl
because pillar rendering happens on master with minion injected opts, we
need to do some special handling to correctly/safely handle it with
existing acl. This patch uses the 'default' acl if set, since there is
no RequestContext actually available during a pillar render. It is
hypothetically possible we could propogate this from minion to master
for a given request in payload via auth_check, but to do so would trust
the minion context, which we generally do not since arbitrary execution
is possible.