new feature: pillar_acl_module
we use this to generate the acl specifically for when pillar is
rendered. the reason this is needed is due to the chicken/egg problem of
the other eauth_acl_module generation being based off of pillar itself;
we fetch just the tenancies a node is associated with; and check if the
equivalent role account exists in external_auth['never'][$tenancy]. this
isn't perfect but will do for now.

NOTE: salt.auth.environments is blp specific and not meant for upstream